finito crud

+++ app/Http/Controllers/admin/ProjectController.php
@@ -92,7 +92,9 @@ class ProjectController extends Controller
         $project->slug = Str::slug($data['name']);
 
         if (isset($data['cover_image'])) {
-            Storage::disk('public')->delete($project->cover_image);
+            if ($project->cover_image) {
+                Storage::disk('public')->delete($project->cover_image);
+            }
             $data['cover_image'] = Storage::disk('public')->put('uploads', $data['cover_image']);
         }
 
@@ -108,6 +110,11 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        //
+        if ($project->cover_image) {
+            Storage::disk('public')->delete($project->cover_image);
+        }
+
+        $project->delete();
+        return redirect()->route('admin.projects.index');
     }
 }
